Explicit cast to (uint8_t) for storage class
The (char)class variable is always either:
  'c': chunk
  'i': metaindex
  'm': metadata
and these values are safe to cast to (uint8_t).

+++ tar/storage/storage_directory.c
@@ -54,7 +54,7 @@ storage_directory_read(uint64_t machinenum, char class, int key,
 	/* Initialize structure. */
 	C.machinenum = machinenum;
 	C.done = 0;
-	C.class = class;
+	C.class = (uint8_t)class;
 	C.key = key;
 	memset(C.start, 0, 32);
 	C.flist = NULL;
